Disable the auto-load functionality
Previously, when RDocumentation was loaded the first time, it would automatically update .RProfile so that next time you start R, RDocumentation would be loaded automatically. This caused confusion and was not accepted by CRAN's guidelines. This commit fixes that: all of the cruft is removed. The help functionality is only overridden if you explicitly load the RDocumentation package.